feat: setup screen flow
This commit introduces an unconventional way of displaying screens in a
given order. The solution uses a FutureBuilder where the future is an
asynchronous method with multiple calls to Navigator.push() which are
waited for. I have done my best to test it and I do not think it will
have any side effects but if we in the future discover issues related
to the setup this commit might be a good place to start the
investigation.